+++ go.mod
@@ -49,7 +49,6 @@ require (
 
 require (
 	github.com/alessio/shellescape v1.4.1
-	github.com/asaskevich/govalidator v0.0.0-20230301143203-a9d515a09cc2
 	github.com/docker/distribution v2.8.1+incompatible
 	github.com/joho/godotenv v1.4.0
 	github.com/joshdk/go-junit v1.0.0
