+++ url/git_url_test.go
@@ -17,9 +17,6 @@ limitations under the License.
 package url
 
 import (
-	"fmt"
-	"testing"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"knative.dev/pkg/apis"
@@ -53,62 +50,3 @@ var _ = Describe("Test.MatchGitURLPrefix", func() {
 		Entry("test suffix /", "https://github.com/katanomi/pkg", "https://github.com/katanomi/pkg.git", true),
 	)
 })
-
-func TestExpandURLIPv6(t *testing.T) {
-	tests := map[string]struct {
-		rawURL  string
-		want    string
-		wantErr error
-	}{
-		"ipv4 url": {
-			rawURL: "http://172.26.168.90",
-			want:   "http://172.26.168.90",
-		},
-		"ipv4 port url": {
-			rawURL: "http://172.26.168.90:8080",
-			want:   "http://172.26.168.90:8080",
-		},
-		"ipv6 url": {
-			rawURL: "http://[20::172:26:168:90]",
-			want:   "http://[0020:0000:0000:0000:0172:0026:0168:0090]",
-		},
-		"ipv6 port url": {
-			rawURL: "http://[20::172:26:168:90]:8080",
-			want:   "http://[0020:0000:0000:0000:0172:0026:0168:0090]:8080",
-		},
-		"ipv6 with ipv4 url": {
-			rawURL: "http://[::FFFF:192.168.0.1]",
-			want:   "http://[0000:0000:0000:0000:0000:ffff:c0a8:0001]",
-		},
-		"ipv6 with ipv4 port url": {
-			rawURL: "http://[::FFFF:192.168.0.1]:8080",
-			want:   "http://[0000:0000:0000:0000:0000:ffff:c0a8:0001]:8080",
-		},
-		"domain url": {
-			rawURL: "http://test.com",
-			want:   "http://test.com",
-		},
-		"domain port url": {
-			rawURL: "http://test.com:8080",
-			want:   "http://test.com:8080",
-		},
-		"paser failed": {
-			rawURL:  "http\n://[20::172::26:168:90]:8080",
-			want:    "",
-			wantErr: fmt.Errorf("net/url: invalid control character in URL"),
-		},
-	}
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			g := NewGomegaWithT(t)
-			got, err := ExpandURLIPv6(tt.rawURL)
-			if tt.wantErr != nil {
-				g.Expect(err).NotTo(BeNil())
-				g.Expect(err.Error()).To(ContainSubstring(tt.wantErr.Error()))
-			} else {
-				g.Expect(err).To(BeNil())
-			}
-			g.Expect(got).To(Equal(tt.want), name)
-		})
-	}
-}
